Handle incomplete writev() operations
The writev() call is allowed to write fewer bytes than requested.  This
is now handled properly to complete the write.  It is not a fully robust
solution, but one that will work for most cases.  Cases where it doesn't
work can be avoided by using a different (usually smaller) chunk size.